+++ index.js
@@ -15,8 +15,20 @@ const postApproved = require('./triggers/post_approved');
 const postFailed = require('./triggers/post_failed');
 const postPublished = require('./triggers/post_published');
 const postRejected = require('./triggers/post_rejected');
+const listPinterestBoards = require('./triggers/list_pinterest_boards');
+const listTikTokPrivacyLevels = require('./triggers/list_tiktok_privacy_levels');
+const {
+  listYouTubeChannels,
+  listTikTokChannels,
+  listPinterestChannels,
+  listLinkedInChannels,
+} = require('./triggers/platform_channel_lists');
 const approvePost = require('./creates/approve_post');
 const createPost = require('./creates/create_post');
+const createYouTubeVideo = require('./creates/create_youtube_video');
+const createTikTokPost = require('./creates/create_tiktok_post');
+const createPinterestPin = require('./creates/create_pinterest_pin');
+const createLinkedInPost = require('./creates/create_linkedin_post');
 const deletePost = require('./creates/delete_post');
 const publishPost = require('./creates/publish_post');
 const rejectPost = require('./creates/reject_post');
@@ -53,11 +65,21 @@ module.exports = {
     [postFailed.key]: postFailed,
     [postPublished.key]: postPublished,
     [postRejected.key]: postRejected,
+    [listYouTubeChannels.key]: listYouTubeChannels,
+    [listTikTokChannels.key]: listTikTokChannels,
+    [listPinterestChannels.key]: listPinterestChannels,
+    [listLinkedInChannels.key]: listLinkedInChannels,
+    [listPinterestBoards.key]: listPinterestBoards,
+    [listTikTokPrivacyLevels.key]: listTikTokPrivacyLevels,
   },
 
   creates: {
     [approvePost.key]: approvePost,
     [createPost.key]: createPost,
+    [createYouTubeVideo.key]: createYouTubeVideo,
+    [createTikTokPost.key]: createTikTokPost,
+    [createPinterestPin.key]: createPinterestPin,
+    [createLinkedInPost.key]: createLinkedInPost,
     [deletePost.key]: deletePost,
     [publishPost.key]: publishPost,
     [rejectPost.key]: rejectPost,